Redid capstone part 1/2 of it!
Where it can handle: The engine should accept arbitrarily large inputs

+++ src/numberlettercountdatastoring/DataStoreApi.java
@@ -1,13 +1,15 @@
 package numberlettercountdatastoring;
+
+import java.math.BigInteger;
 import java.util.List;
 
 import project.annotations.ProcessAPI;
 
 @ProcessAPI
 public interface DataStoreApi {
-	int insertRequest(DataRequest dataRequest);    
+	int insertRequest(DataRequest dataRequest);
 	boolean validateNumber(int number);
 	boolean writeResultsToFile(String filePath, List<String> results);
-	// Read numbers from an input file and return the list of integers found
-	List<Integer> processFile(String input);
+	// Read numbers from an input file and return the list of BigInteger values found
+	List<BigInteger> processFile(String input);
 }
